gen: Extrace AWSRequest generation into its own module
